mount_shoes/requirements.txt
django==3.2.12
djangorestframework==3.13.1
django-environ==0.8.1
pillow==9.0.1
django-extensions==3.1.5
werkzeug==2.0.3
mysqlclient==2.1.0
django_extensions==3.1.5
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
django-cors-headers==3.11.0
paypalrestsdk==1.13.1